split python code
