+++ jupyter_server_nbmodel/__init__.py
@@ -8,9 +8,9 @@ except ImportError:
     # in editable mode with pip. It is highly recommended to install
     # the package from a stable release or in editable mode: https://pip.pypa.io/en/stable/topics/local-project-installs/#editable-installs
     import warnings
-
     warnings.warn("Importing 'jupyter_server_nbmodel' outside a proper installation.", stacklevel=1)
     __version__ = "dev"
+
 from .extension import Extension
