+++ jupyter_server_nbmodel/event_logger.py
@@ -1,8 +1,10 @@
 # Copyright (c) 2024-2025 Datalayer, Inc.
 # Distributed under the terms of the Modified BSD License.
 
-from jupyter_events import EventLogger
 import pathlib
+
+from jupyter_events import EventLogger
+
 
 _JUPYTER_SERVER_EVENTS_URI = "https://events.jupyter.org/jupyter_server_nbmodel"
