+++ jupyter_server_nbmodel/extension.py
@@ -8,8 +8,13 @@ import asyncio
 from jupyter_server.extension.application import ExtensionApp
 from jupyter_server.services.kernels.handlers import _kernel_id_regex
 
-from .handlers import ExecuteHandler, ExecutionStack, InputHandler, RequestHandler
-from .log import get_logger
+from jupyter_server_nbmodel.execution_stack import ExecutionStack
+from jupyter_server_nbmodel.handlers import (
+    ExecuteHandler,
+    InputHandler,
+    RequestHandler,
+)
+from jupyter_server_nbmodel.log import get_logger
 
 
 RTC_EXTENSIONAPP_NAME = "jupyter_server_ydoc"
@@ -20,7 +25,9 @@ REQUEST_ID_REGEX = r"(?P<request_id>\w+-\w+-\w+-\w+-\w+)"
 
 
 class Extension(ExtensionApp):
+
     name = "jupyter_server_nbmodel"
+
 
     def initialize_handlers(self):
         rtc_extension = None
@@ -32,11 +39,10 @@ class Extension(ExtensionApp):
             if n_extensions > 1:
                 get_logger().warning("%i collaboration extensions found.", n_extensions)
             rtc_extension = next(iter(rtc_extensions))
-
         self.__execution_stack = ExecutionStack(
-            manager=self.settings["kernel_manager"], ydoc_extension=rtc_extension
+            manager=self.settings["kernel_manager"],
+            ydoc_extension=rtc_extension,
         )
-
         self.handlers.extend(
             [
                 (
@@ -56,6 +62,7 @@ class Extension(ExtensionApp):
                 ),
             ]
         )
+
 
     async def stop_extension(self):
         if hasattr(self, "__execution_stack"):
